KTOR-6108 KtorServlet does not support yaml configuration
