+++ ktor-server/ktor-server-servlet-jakarta/jvm/test-resources/application.yaml
@@ -1,0 +1,4 @@
+ktor:
+    deployment:
+        port: 1234
+property: a
