+++ ktor-server/ktor-server-servlet/build.gradle.kts
@@ -12,6 +12,7 @@ kotlin.sourceSets {
     jvmTest {
         dependencies {
             api(project(":ktor-server:ktor-server-core", configuration = "testOutput"))
+            api(project(":ktor-server:ktor-server-config-yaml"))
             implementation(libs.mockk)
             implementation(libs.javax.servlet)
         }
