+++ ktor-server/ktor-server-servlet/jvm/test-resources/custom-config.yaml
@@ -1,0 +1,4 @@
+ktor:
+    deployment:
+        port: 1234
+property: a-custom
